Support #hasOwnProperty in content proxies
`#hasOwnProperty` is used by nunjucks in trying to decide if the last
argument in a macro call is a dict of keywords or a object that should
be treated as a single argument.

Without this change a macro called with content in the last arg
position would throw an error.