Allow steps to declare that the user should stop here
This is less important atm but in the future the function of goTo will
change to automatically jump a user to the next incomplete step. When
that happens each step will need a way to declare:

- despite this step being complete you should still stop here
  (check your answers for example)
- if leaving this step you should always redirect to the next step, not
  auto jumping (address postcode to address picker step)

Stop provides for the first use case, a new form of Redirector will be
needed to satisfy the second.