Add the ability to force Subsequent questions even if they have already been answered
Sometimes you have a question that has a number of follow on questions.
If you change your response to the first question you may wish to force
the user to re-answer the subsequent questions.

To do this you need to use the ifCompleteAndNotForceShowThenContinue
flow control by overriding flowControl in your question. You then
need to set this.journey.forceShow to be an array of the questions you
want to be re-answered.

This logic has been combined into QuestionWithRequiredNextSteps which
can be extended in place of Question. When you extend it you must
implement a requiredNextSteps method which returns an array of steps
you want to be re-answered.